README: Update devicetree overlays
Now kernels (at least as of Debian linux-image-6.5.0-2-rpi) contain
the symbols necessary for clean overlay creation, removing the
need for the hackish &{/soc} section. Unfortunately it still does
not remove the need for #*-cells duplication (or I could not figure
out how to avoid it).
Also, fix the i2c bus symbol, it is i2c1.